GenAI_Krish/simple_genAI/requirements.txt
langchain
python-dotenv
pymupdf
langchain-community
langchain-text-splitters
langchain-openai
faiss-cpu
langchain-classic